book_notes_app/requirements.txt
# 基础依赖
PyQt5==5.15.9
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.8.0.76

# OCR相关
paddlepaddle==2.5.1
paddleocr==2.7.0

# 工具库
requests==2.31.0
tqdm==4.66.1
colorama==0.4.6
python-dotenv==1.0.0

# 图像处理
scikit-image==0.21.0

